 requirements.txt
torch==2.1.0
torchvision==0.15.0
scikit-learn==1.3.0
Pillow==9.5.0
matplotlib==3.8.0
pandas==2.1.0